Change background color to 'light' again
See: https://www.quora.com/Which-is-better-for-the-eyes-a-white-background-with-black-text-a-black-background-with-white-text-or-sepia-mode

Most studies have shown that reading digital books written with black
characters on white (light) background is better for the eyes than
reading the same in white text with black (dark) background. In the
experiments, the subjects read with 26% greater accuracy when they read
with dark characters on a light/white background. The science (physics)
behind this phenomenon is the following:

1) When the background is white, display is bright and the iris closes a
bit more than normal and this decreases the effect of the "deformed"
lens. With a black background the iris opens a bit more to receive more
light and the deformation of the lens creates a much fuzzier focus at
the eye. This is known as "fuzzing effect" or "Halation".

2) While reading an e-book, white stimulates, in nearly equal amount,
all the 3 colour receptors in our eyes . This makes reading white text
on dark background stressful on the eyes.

3)It is easy to read a digital book written in black text on white
background because the light that makes us read each word and letter is
not reflected but absorbed. This causes much less strain on our eyes. In
the case of black background with white text, the effect is opposite
because the reflected light from text scatter into each other. Addition
of grey tint to the background improves the situation because it means
less light reflects behind the words, making it easier on the eyes.

4) Black text works better because black is a colour that does not
reflect light in any part of the visible spectrum.

5) White reflects all wavelengths of light. The characters in the book
being compact and close, when white reading material reflects light, the
reflected light scatters and runs into neighbouring words and letters.
This makes the shape of the text material harder to perceive and this
strains the eyes. In comparison, when the text is black, the black
colour absorbs the light around each word and letter making them easy to
distinguish from one another. This causes less strains on the eyes.

6)With a tint of grey in the background, less light reflects behind the
words, making it easier on the eyes. Since black colour does not reflect
light in any part of the visible spectrum, the black text works better.
Therefore, fixating on black text while reading will not put as much
stress on the readers' eyes because it absorbs the light that strikes
each word.

7)If you use a dark background, the text should be displayed in grey
tint. This will not put as much stress on our eyes because grey text is
not as bright as white text. It reflects less light making it easier to
read. In a dark room, white text on a black ground is not much stressful
on the eyes. This is because no light is reflecting off it in a dark
room.

8) Black text on a light-grey background is easier to read than in a
white background because less light is reflected behind the text. Grey
text on dark background is easier to read than white text on dark
background because less light is reflected on words.

9) While scanning text, white text on dark background is fine. White
reflects all the colours of the visible light spectrum into the eyes.
This makes the text look bright and distinct. Since scanning of a
heading, a title or a label takes a quick sweep of the eyes, there is no
worry as far as putting stress on the readers' eyes is concerned.

10)Black letters on white background can be bad for people with dyslexia.

11)Most readable colour combination is black text on white background.

12)Computer screens being projections, the black-white contrast is more
pronounced than it is on a printed book. This increased contrast causes
eye-strain quickly.

13)How the text material is placed on the page of a book is also
important. If the reading material is embedded and/or part of graphic,
size of pixel matters.

14)Nearly 50% of population have astigmatism and these people find it
harder to read white alpha-numeric characters on black than black on
white.

SEPIA MODE OF READING DIGITAL BOOKS: Research shows that switching from
white to sepia background, the strain on the eye reduces since sepia
spectrum is softer than the white spectrum. The effective radiance in
sepia background is nearly 25% lower than in white background. This
means lower screen luminescence and hence lesser visual fatigue, lesser
strain on the eyes. Visual eye fatigue is lesser while reading in sepia
background as compared to white background. There is more eye strain,
blurry vision and tiredness in white background in comparison with sepia
background.